refactor: move backcore types in the libs
