+++ apps/drec-api/src/lib/backend-core/utils/devices.ts
