+++ apps/drec-api/src/lib/backend-core/utils/email-confirmation.ts
